+++ rivertion/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.1.0-dev.2
+- Now the errors of onFailed, onSuccess, onSettled are brought back to BlocObserver
+- Fix DebounceBloc not correct emits state 
+
 ## 0.1.0-dev.1
 
 - Perform asynchronous data mutations